Update comments.json
translation of “read” to “leitura” within the post page just below the title displays the date of publication and an estimate of how long it will take to read the post, there is no translation into Brazilian Portuguese. The suggestion I'm making is to translate the word “read” in English into “leitura” in Brazilian Portuguese.